+++ lib/app/modules/announcement/announcement_list_controller.dart
@@ -127,35 +127,6 @@ class AnnouncementListController extends GetxController {
     searchedAnnouncements.value = announcements.value;
   }
 
-  Future<void> toggleBookmark(String postId) async {
-    if (announcements.value == null) return;
-
-    final index = announcements.value!.results.indexWhere(
-      (a) => a.postId == postId,
-    );
-
-    final request = await EasipRouter.putBookmark(houseId: postId);
-    final response = await _dataSource.execute(request);
-
-    // API 응답이 성공일 때만 상태 업데이트
-    if (response?.success ?? false) {
-      // 원본 데이터 업데이트
-      announcements.value!.results[index].isPushAlarmRegistered.value = 
-          !announcements.value!.results[index].isPushAlarmRegistered.value;
-
-      // 검색 결과가 있다면 동일한 공고 업데이트
-      if (searchedAnnouncements.value != null) {
-        final searchIndex = searchedAnnouncements.value!.results.indexWhere(
-          (a) => a.postId == postId,
-        );
-        if (searchIndex != -1) {
-          searchedAnnouncements.value!.results[searchIndex].isPushAlarmRegistered.value = 
-              announcements.value!.results[index].isPushAlarmRegistered.value;
-        }
-      }
-    }
-  }
-
   Future<void> getNextAnnouncements() async {
     if (isLoading.value || searchedAnnouncements.value == null) return;
