+++ lib/app/modules/announcement/announcement_list_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'announcement_list_controller.dart';
-import '../../core/utils/screen_utils.dart';
-import 'model/announcement_response.dart';
+import 'announcement_row.dart';
+
 
 class AnnouncementListView extends GetView<AnnouncementListController> {
   const AnnouncementListView({super.key});
@@ -128,7 +128,6 @@ class AnnouncementListView extends GetView<AnnouncementListController> {
                           
                           return AnnouncementRow(
                             announcement: announcement,
-                            onBookmarkToggle: () => controller.toggleBookmark(announcement.postId),
                           );
                         },
                       );
@@ -136,121 +135,6 @@ class AnnouncementListView extends GetView<AnnouncementListController> {
                   ),
                 ),
               ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class AnnouncementRow extends StatelessWidget {
-  final Announcement announcement;
-  final VoidCallback onBookmarkToggle;
-
-  const AnnouncementRow({
-    super.key,
-    required this.announcement,
-    required this.onBookmarkToggle,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.only(left: 16, right: 16, top: 8),
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Stack(
-            children: [
-              SizedBox(
-                width: ScreenUtils.ratioWidth(context, 75),
-                height: ScreenUtils.ratioWidth(context, 75),
-                child: ClipRRect(
-                  borderRadius: BorderRadius.circular(4),
-                  child: announcement.hasImage
-                      ? Image.network(
-                          announcement.thumbnailUrl,
-                          fit: BoxFit.cover,
-                          errorBuilder: (context, error, stackTrace) {
-                            return Container(
-                              color: Colors.grey[300],
-                              child: Icon(
-                                Icons.business,
-                                size: 30,
-                                color: Colors.grey[600],
-                              ),
-                            );
-                          },
-                        )
-                      : Container(
-                          color: Colors.grey[300],
-                          child: Icon(
-                            Icons.business,
-                            size: 30,
-                            color: Colors.grey[600],
-                          ),
-                        ),
-                ),
-              ),
-              Positioned(
-                top: 0,
-                left: 0,
-                child: Container(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 6,
-                    vertical: 2,
-                  ),
-                  // decoration: BoxDecoration(
-                  //   color: announcement.isActive ? Colors.blue : Colors.grey,
-                  //   borderRadius: BorderRadius.circular(4),
-                  // ),
-                  child: Text(
-                    announcement.subscriptionState,
-                    style: const TextStyle(
-                      color: Colors.white,
-                      fontSize: 10,
-                      fontWeight: FontWeight.w500,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(width: 12),
-          Expanded(
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  announcement.title,
-                  style: const TextStyle(
-                    fontSize: 16,
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-                const SizedBox(height: 4),
-                Text(
-                  announcement.periodText,
-                  style: TextStyle(fontSize: 14, color: Colors.grey[600]),
-                ),
-                Text(
-                  announcement.recruitingText,
-                  style: TextStyle(fontSize: 14, color: Colors.grey[600]),
-                ),
-              ],
-            ),
-          ),
-          Obx(
-            () => IconButton(
-              onPressed: onBookmarkToggle,
-              icon: Icon(
-                Icons.bookmark_border,
-                color:
-                    announcement.isPushAlarmRegistered.value
-                        ? Colors.red
-                        : Colors.grey[400],
-              ),
             ),
           ),
         ],
