--- lib/app/modules/announcement/model/bookmark_response.dart
@@ -1,17 +1,0 @@
-class BookmarkResponse {
-  final bool success;
-
-  BookmarkResponse({required this.success});
-
-  factory BookmarkResponse.fromJson(Map<String, dynamic> json) {
-    return BookmarkResponse(
-      success: json['success'] ?? false,
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'success': success,
-    };
-  }
-}
